Mosaic with image on top
I added an image on top of my collage, played around with the saturation, contrast and brightness and now have this. I would say I'm more than half way done with it, but there are a couple more things I want to play around with before I turn in my final.